Make stdPeopleMessage protected
There are a number of "people" operations that are not supported through
the Java API (such as `$unset`). Currently users are forced to
reimplement (read: copy/paste) the logic in stdPeopleMessage when they
should be able to add a simple one-line method that calls through to
stdPeopleMessage with a specific `actionType`.